requirements.txt
selenium==3.141.0
python-dotenv==0.21.0
urllib3>=1.26.15,<2.0
discord.py==1.7.2